Changement de type d'élément dans la pile, maintenant des int

+++ include/elementAudio.h
@@ -1,12 +1,12 @@
 #ifndef ELEMENT_H
 #define ELEMENT_H
 typedef struct ELEMENT{
-	double val;
+	int val;
 }ELEMENT;
 
 void affiche_ELEMENT(ELEMENT display);
 ELEMENT saisir_ELEMENT();
-ELEMENT affect_ELEMENT(double value);
+ELEMENT affect_ELEMENT(int value);
 int compare_ELEMENT(ELEMENT a, ELEMENT b);
 
 #endif
